test: clean support for both compiler and parser fixtures
