+++ packages/@lwc/template-compiler/src/__tests__/fixtures.spec.ts
@@ -9,24 +9,31 @@ import path from 'path';
 import glob from 'glob';
 import prettier from 'prettier';
 
-import compiler from '../index';
+import compiler, { parse } from '../index';
 
-const FIXTURE_DIR = path.join(__dirname, 'fixtures');
-const COMPILER_FIXTURE_DIR = path.join(FIXTURE_DIR, 'compiler');
+const FIXTURES_DIR = path.join(__dirname, 'fixtures');
 const BASE_CONFIG = {};
 
 const EXPECTED_JS_FILENAME = 'expected.js';
+const EXPECTED_JSON_FILENAME = 'expected.json';
 const EXPECTED_META_FILENAME = 'metadata.json';
 
 const ONLY_FILENAME = '.only';
 const SKIP_FILENAME = '.skip';
 
 describe('fixtures', () => {
-    const fixtures = glob.sync(path.resolve(COMPILER_FIXTURE_DIR, '**/*.html'));
+    const fixtures = glob.sync(path.resolve(FIXTURES_DIR, '**/*.html'));
 
     for (const caseEntry of fixtures) {
         const caseFolder = path.dirname(caseEntry);
-        const caseName = path.relative(COMPILER_FIXTURE_DIR, caseFolder);
+        const caseName = path.relative(FIXTURES_DIR, caseFolder);
+
+        const category = caseName.split(path.sep).shift();
+        if (category !== 'compiler' && category !== 'parser') {
+            throw new Error(
+                `Unexpected category "${category}" encountered while running fixture test for: ${caseName}`
+            );
+        }
 
         const fixtureFilePath = (fileName): string => {
             return path.join(caseFolder, fileName);
@@ -56,27 +63,48 @@ describe('fixtures', () => {
 
         testFn(`${caseName}`, () => {
             const src = readFixtureFile('actual.html');
-
             const configOverride = JSON.parse(readFixtureFile('config.json'));
-            let expectedCode = readFixtureFile(EXPECTED_JS_FILENAME);
-            let expectedMetaData = JSON.parse(readFixtureFile(EXPECTED_META_FILENAME));
 
-            const actual = compiler(src, {
-                ...BASE_CONFIG,
-                ...configOverride,
-            });
+            let actual;
 
-            if (expectedCode === null) {
-                // write compiled js file if doesn't exist (ie new fixture)
-                expectedCode = actual.code;
-                writeFixtureFile(
-                    EXPECTED_JS_FILENAME,
-                    prettier.format(expectedCode, {
-                        parser: 'babel',
-                    })
+            if (category === 'compiler') {
+                actual = compiler(src, {
+                    ...BASE_CONFIG,
+                    ...configOverride,
+                });
+
+                let expectedCode = readFixtureFile(EXPECTED_JS_FILENAME);
+                if (expectedCode === null) {
+                    // write compiled js file if doesn't exist (ie new fixture)
+                    expectedCode = actual.code;
+                    writeFixtureFile(
+                        EXPECTED_JS_FILENAME,
+                        prettier.format(expectedCode, {
+                            parser: 'babel',
+                        })
+                    );
+                }
+                // check compiled code
+                expect(prettier.format(actual.code, { parser: 'babel' })).toEqual(
+                    prettier.format(expectedCode, { parser: 'babel' })
+                );
+            }
+            if (category === 'parser') {
+                actual = parse(src, configOverride);
+
+                let expectedAST = JSON.parse(readFixtureFile(EXPECTED_JSON_FILENAME));
+                if (expectedAST === null) {
+                    // write file if doesn't exist (ie new fixture)
+                    expectedAST = actual.root;
+                    writeFixtureFile(EXPECTED_JSON_FILENAME, JSON.stringify(expectedAST, null, 4));
+                }
+                // check serialized ast
+                expect(JSON.stringify(actual.root, null, 4)).toEqual(
+                    JSON.stringify(expectedAST, null, 4)
                 );
             }
 
+            let expectedMetaData = JSON.parse(readFixtureFile(EXPECTED_META_FILENAME));
             if (expectedMetaData === null) {
                 // write metadata file if doesn't exist (ie new fixture)
                 const metadata = {
@@ -85,13 +113,8 @@ describe('fixtures', () => {
                 expectedMetaData = metadata;
                 writeFixtureFile(EXPECTED_META_FILENAME, JSON.stringify(expectedMetaData, null, 4));
             }
-
             // check warnings
             expect(actual.warnings).toEqual(expectedMetaData.warnings || []);
-            // check compiled code
-            expect(prettier.format(actual.code, { parser: 'babel' })).toEqual(
-                prettier.format(expectedCode, { parser: 'babel' })
-            );
         });
     }
 });
