+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parser/component-with-data-binding/config.json
