+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parser/component-without-data-binding/metadata.json
