+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parser/for-each-inline/config.json
