+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parser/if-inline/expected.json
