+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parser/if-template/config.json
