+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parser/iterator-inline/expected.json
