+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parser/iterator-inline/metadata.json
